umbot/Backend/requirements.txt
openai
python-dotenv
django 
djangorestframework
requests
